% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/boxplot_RIGHT.R
\name{boxplot_RIGHT}
\alias{boxplot_RIGHT}
\title{Box-whisker}
\usage{
boxplot_RIGHT(form, data, col = NULL, isString = FALSE)
}
\arguments{
\item{form}{a formula describing the x and y variables as y ~ x.}

\item{data}{a data.frame object.}

\item{col}{color of the boxes.}

\item{isString}{a character is expected for \code{x} and \code{data} if \code{TRUE}. It is useful for programming.}
}
\description{
Draw a box-whisker of the given data values.
}
\examples{
\dontrun{
RIGHT(boxplot(conc ~ Subject, Theoph))
}
}
\seealso{
\code{\link{boxplot}}
}

