% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/qplot_RIGHT.R
\name{createQplot}
\alias{createQplot}
\title{Extract data object's name}
\usage{
createQplot(x, y = NULL, ..., data, geom = "point")
}
\arguments{
\item{x}{xaxis to draw graph.}

\item{y}{yaxis to draw graph. Default is NULL}

\item{...}{other options to draw graph(colour, fill)}

\item{data}{a data.frame object.}

\item{geom}{graph type to draw(point, line, bar, boxplot). Default is "point"}
}
\description{
Function to create data object name in qplot function.
}
\examples{
\dontrun{
RIGHT({qplot(x=Time, y=conc, data=Theoph, geom="point", colour=Subject)
       qplot(x=Time, fill=Subject, data=Theoph, geom="bar")})
}
}
\seealso{
\code{\link{qplot}}
}

