% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ggplot_RIGHT.R
\name{ggplot_RIGHT}
\alias{ggplot_RIGHT}
\title{Make RIGHT html code using ggplot object}
\usage{
ggplot_RIGHT(obj)
}
\arguments{
\item{obj}{a data.frame object.}
}
\description{
Function to create data object name in ggplot function.
}
\examples{
\donttest{
obj <- ggplot(Theoph, aes(Time, conc)) + geom_point()
attr(obj, "NAME") <- "Theoph"
RIGHT(print(obj))
}
}
\seealso{
\code{\link{ggplot2}}
}

