% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/points_RIGHT.R
\name{points_RIGHT}
\alias{points_RIGHT}
\title{Add Points to a Plot}
\usage{
points_RIGHT(form, data, isString = FALSE)
}
\arguments{
\item{form}{a formula describing the x and y variables as y ~ x.}

\item{data}{a data.frame object.}

\item{isString}{a character is expected for \code{data} if \code{TRUE}. It is useful for programming.}
}
\description{
points_RIGHT draws a sequence of points in the current axis.
}
\examples{
\dontrun{
RIGHT({plot(conc ~ Time, Theoph, type = "n")
       points(conc ~ Time, Theoph)})
}
}
\seealso{
\code{\link{points}}
}

