% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RIGHT.R
\name{print.RIGHT}
\alias{print.RIGHT}
\title{Print RIGHT Object}
\usage{
\method{print}{RIGHT}(x, ...)
}
\arguments{
\item{x}{RIGHT object.}

\item{...}{additional arguments affecting the summary produced.}
}
\description{
Print RIGHT Object
}
\examples{
\dontrun{
obj <- RIGHT(plot(conc ~ Time, Theoph), Theoph)
print(obj)
}
}

