% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/runServer.R
\name{runServer}
\alias{runServer}
\title{shiny server-offloading}
\usage{
runServer(expr = { })
}
\arguments{
\item{expr}{a formula to draw server-offloading graph and return result of formula.}
}
\description{
Function to create server-offloading graph.
}
\examples{
\dontrun{
obj <-  RIGHT({plot(conc ~ Time, Theoph, type="p", color = Subject)
               loessArray <- runServer({obj <- loess(conc ~ Time, data = Theoph)
                   xRange <- range(Theoph$Time)
                   simArray <- data.frame(Time = seq(xRange[1], xRange[2], length.out = 132))
                   simArray$conc <- predict(obj, newdata = simArray)
                   return(simArray)})
               lines(conc ~ Time, loessArray)})
print(obj)
}
}
\seealso{
\code{\link{shiny}}
}

