% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/search_RIGHT.R
\name{search_RIGHT}
\alias{search_RIGHT}
\title{Add a Search Box}
\usage{
search_RIGHT(data, isString = FALSE)
}
\arguments{
\item{data}{a data.frame object.}

\item{isString}{a character is expected for \code{data} if \code{TRUE}. It is useful for programming.}
}
\description{
search_RIGHT adds a search box for a \code{data.frame} object.
}
\examples{
\dontrun{
RIGHT({plot(conc ~ Time, Theoph, type = "p", color = Subject)
       search(Theoph)})
}
}

