% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RIGHT.R
\name{summary.RIGHT}
\alias{summary.RIGHT}
\title{Summarize RIGHT Object}
\usage{
\method{summary}{RIGHT}(object, ...)
}
\arguments{
\item{object}{RIGHT object.}

\item{...}{additional arguments affecting the summary produced.}
}
\description{
Summarize RIGHT Object
}
\examples{
\dontrun{
obj <- RIGHT(plot(conc ~ Time, Theoph), Theoph)
summary(obj)
}
}

