% Generated by roxygen2 (4.0.1): do not edit by hand
\name{table_RIGHT}
\alias{table_RIGHT}
\title{Add an Interactive Table}
\usage{
table_RIGHT(data, height = 200L, isString = FALSE)
}
\arguments{
\item{data}{a data.frame object.}

\item{height}{height of the table. The default is 200.}

\item{isString}{a character is expected for \code{data} if \code{TRUE}. It is useful for programming.}
}
\description{
table_RIGHT adds an interactive table for a \code{data.frame} object.
}
\examples{
\donttest{
obj <- RIGHT({plot(conc ~ Time, Theoph, type = "p", color = Subject)
              table(Theoph)})
print(obj)
}
}

