% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/lines_RIGHT.R
\name{lines_RIGHT}
\alias{lines_RIGHT}
\title{Add Lines to a Plot}
\usage{
lines_RIGHT(form, data, by = NULL, isString = FALSE)
}
\arguments{
\item{form}{a formula describing the x and y variables as y ~ x.}

\item{data}{a data.frame object.}

\item{by}{column used to group lines. Default is NULL.}

\item{isString}{a character is expected for \code{data} if \code{TRUE}. It is useful for programming.}
}
\description{
lines_RIGHT draws a sequence of points in the current axis.
}
\examples{
\dontrun{
RIGHT({plot(conc ~ Time, Theoph, type = "n")
       lines(conc ~ Time, Theoph)})
}
}
\seealso{
\code{\link{lines}}
}

