% Generated by roxygen2 (4.0.1): do not edit by hand
\name{pie_RIGHT}
\alias{pie_RIGHT}
\title{Pie Charts}
\usage{
pie_RIGHT(x, data, isString = FALSE)
}
\arguments{
\item{x}{variable name for which the histogram is desired.}

\item{data}{a data.frame object.}

\item{isString}{a character is expected for \code{x} and \code{data} if \code{TRUE}. It is useful for programming.}
}
\description{
Draw a pie chart.
}
\examples{
\donttest{
obj <- RIGHT(pie(Subject, Theoph))
print(obj)
}
}
\seealso{
\code{\link{pie}}
}

