% Generated by roxygen2 (4.0.1): do not edit by hand
\name{points_RIGHT}
\alias{points_RIGHT}
\title{Add Points to a Plot}
\usage{
points_RIGHT(form, data, isString = FALSE)
}
\arguments{
\item{form}{a formula describing the x and y variables as y ~ x.}

\item{data}{a data.frame object.}

\item{isString}{a character is expected for \code{data} if \code{TRUE}. It is useful for programming.}
}
\description{
points_RIGHT draws a sequence of points in the current axis.
}
\examples{
\donttest{
obj <- RIGHT({plot(conc ~ Time, Theoph, type = "n") # create blank axis
              points(conc ~ Time, Theoph)})
print(obj)
}
}
\seealso{
\code{\link{points}}
}

