% Generated by roxygen2 (4.0.1): do not edit by hand
\name{runServer}
\alias{runServer}
\title{shiny server-offloading}
\usage{
runServer(expr = { })
}
\arguments{
\item{expr}{a formula to draw server-offloading graph and return result of formula.}
}
\description{
Function to create server-offloading graph.
}
\examples{
\donttest{
obj <-  RIGHT({plot(conc ~ Time, Theoph, type = "p", color = Subject)
               loessArray <- runServer({
                 obj <- loess(conc ~ Time, Theoph)
                 simArray <- data.frame(Time = seq(min(Theoph$Time), max(Theoph$Time), length.out = 100))
                 simArray$conc <- predict(obj, simArray)
                 return(simArray)
               })
               lines(conc ~ Time, loessArray)})
print(obj)
}
}
\seealso{
\code{\link{shiny}}
}

