% Generated by roxygen2 (4.0.1): do not edit by hand
\name{summary.RIGHT}
\alias{summary.RIGHT}
\title{Summarize RIGHT Object}
\usage{
\method{summary}{RIGHT}(object, ...)
}
\arguments{
\item{object}{RIGHT object.}

\item{...}{additional arguments affecting the summary produced.}
}
\description{
Summarize RIGHT Object
}
\examples{
\donttest{obj <- RIGHT(plot(conc ~ Time, Theoph), Theoph)}
\donttest{summary(obj)}
}

